packages/audio/gamepad-synth/tools/tts/pyproject.toml
[project]
name = "gamepad-synth-tts"
version = "0.1.0"
description = "Offline TTS clip generator for gamepad-synth voicing announcements"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.55",
]
